Задание 1.18
