+++ src/view/filter-view.js
@@ -1,7 +1,8 @@
 import { createElement } from '../render.js';
 
 function createFilterTemplate() {
-  return `
+  return (
+    `
   <form class="trip-filters" action="#" method="get">
   <div class="trip-filters__filter">
     <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
@@ -21,7 +22,8 @@ function createFilterTemplate() {
   </div>
   <button class="visually-hidden" type="submit">Accept filter</button>
 </form>
-`;
+`
+  );
 }
 
 export default class FilterView {
