+++ src/view/sort-view.js
@@ -1,7 +1,8 @@
 import { createElement } from '../render.js';
 
 function createSortTemplate() {
-  return `
+  return (
+    `
   <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
   <div class="trip-sort__item  trip-sort__item--day">
     <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day">
@@ -28,7 +29,8 @@ function createSortTemplate() {
     <label class="trip-sort__btn" for="sort-offer">Offers</label>
   </div>
 </form>
-`;
+`
+  );
 }
 
 export default class SortView {
